backend/requirements.txt
sqlalchemy
sqlalchemy_utils
psycopg2
typer
strawberry-graphql[fastapi]